feat(money): integer minor-units migration + planning-feature review fixes
Large-scale change across the whole app.

Money model (spec 019):
- Store/compute all monetary amounts as Long minor units (fixed scale 2),
  never Double; conversion helpers in core:model (money/). Room schema v9
  + Migration_8_9; Firestore dual-write (legacy major + encrypted *Minor,
  read *Minor-first); uniqueHash reconstructs legacy amount string for
  cross-version dedup stability.

Stage-8 planning-feature review fixes (recurring / budgets / debts):
- Debts: debt-payment transaction now updates the account balance and is
  reversed (balance + soft-delete) when the payment/debt is deleted; edit
  no longer duplicates the linked transaction; block total < paid with a
  clear error; surface save/delete failures via Snackbar; payment date
  picker; DebtDao.observeById.
- Recurring: re-anchor DatePicker UTC->local (off-by-one fix); deterministic
  uniqueHash + insertOrIgnore so generation is idempotent / dedup-safe across
  devices; YEARLY preserves Feb 29; first occurrence honours dayOfMonth/
  dayOfWeek; catch-up cap; sensible default day pickers.
- Budgets: exclude already-budgeted categories and surface save errors;
  revive soft-deleted budget instead of orphaning its Firestore doc; pure
  mapper; rounded percent.
- Across: replace raw hex with theme tokens; new localized strings.

Tests: GeneratePendingTransactionsUseCase date math, DebtPaymentRepositoryImpl
balance apply/reverse, BudgetRepositoryImpl revive.

+++ core/database/src/main/java/com/atelbay/money_manager/core/database/entity/AccountEntity.kt
@@ -15,7 +15,7 @@ data class AccountEntity(
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val name: String,
     val currency: String,
-    val balance: Double,
+    val balance: Long,
     val createdAt: Long,
     val remoteId: String? = null,
     val updatedAt: Long = 0,
